# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AbortRequest": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "properties": {
          "session_id": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "title": "AbortRequest",
        "type": "object"
      },
...
      "ApiErrorBody": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "properties": {
          "code": {
            "type": "string"
          },
          "details": true,
          "error": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "details",
          "error"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
...
      "CapabilityDiscoveryResponse": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "description": "Response returned for capability discovery.",
        "properties": {
          "capabilities": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object"
          }
        },
        "required": [
          "capabilities"
        ],
        "title": "CapabilityDiscoveryResponse",
        "type": "object"
      },
...
      "ChatRequest": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "properties": {
          "message": {
            "type": "string"
          },
          "mode": {
            "type": [
              "string",
              "null"
            ]
          },
          "session_id": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "allthecodes Web API",
    "version": "v1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/abort": {
      "post": {
        "operationId": "Abort",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbortRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "per-key:session_id"
      }
    },
...
    "/api/capabilities": {
      "get": {
        "operationId": "Capabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityDiscoveryResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      }
    },
...
    "/api/chat": {
      "post": {
        "operationId": "Chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "per-key:session_id"
      }
    },
...
  }
}